KAFKA-3162 Moved InterceptorCallback to be static private inner class of KafkaProducer
